// src/api/generated/models/exportJobStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type ExportJobStatus = typeof ExportJobStatus[keyof typeof ExportJobStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ExportJobStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
} as const;